Add a tweak for fish shell scripts
In the fish shell, the keyword 'end' is used both as a keyword and for
ending conditional logic.

For example, in the trivial function below, the word `end` is used to
terminate the "if" conditional, and it's also used as a keyword to
indicate the end of the function.

This tweak makes the word the same color in both places. I selected
purple to match the Onedark scheme in Atom.

```fish
function foo

    if not string length $argv
        echo "I got nothing."
    end

end
```